requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
compress-json-python==3.0.0
idna==3.10
requests==2.32.3
urllib3==2.2.3
